pyproject.toml
[tool.poetry]
name = "ufo-perf-tracker"
version = "0.1.0"
description = ""
authors = ["Jany Belluz <jany.belluz@daltonmaag.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
